+++ pkg/apis/configuration/v1/kongclusterplugin_types.go
@@ -34,6 +34,7 @@ import (
 // +kubebuilder:printcolumn:name="Age",type=date,JSONPath=`.metadata.creationTimestamp`,description="Age"
 // +kubebuilder:printcolumn:name="Disabled",type=boolean,JSONPath=`.disabled`,description="Indicates if the plugin is disabled",priority=1
 // +kubebuilder:printcolumn:name="Config",type=string,JSONPath=`.config`,description="Configuration of the plugin",priority=1
+// +kubebuilder:printcolumn:name="Programmed",type=string,JSONPath=`.status.conditions[?(@.type=="Programmed")].status`
 
 // KongClusterPlugin is the Schema for the kongclusterplugins API.
 type KongClusterPlugin struct {
@@ -100,6 +101,16 @@ type KongClusterPluginList struct {
 
 // KongClusterPluginStatus represents the current status of the KongClusterPlugin resource.
 type KongClusterPluginStatus struct {
+	// Conditions describe the current conditions of the KongClusterPluginStatus.
+	//
+	// Known condition types are:
+	//
+	// * "Programmed"
+	//
+	// +listType=map
+	// +listMapKey=type
+	// +kubebuilder:validation:MaxItems=8
+	// +kubebuilder:default={{type: "Programmed", status: "Unknown", reason:"Pending", message:"Waiting for controller", lastTransitionTime: "1970-01-01T00:00:00Z"}}
 	Conditions []metav1.Condition `json:"conditions,omitempty"`
 }