+++ pkg/apis/configuration/v1/kongconsumer_types.go
@@ -29,6 +29,7 @@ import (
 // +kubebuilder:validation:Optional
 // +kubebuilder:printcolumn:name="Username",type=string,JSONPath=`.username`,description="Username of a Kong Consumer"
 // +kubebuilder:printcolumn:name="Age",type=date,JSONPath=`.metadata.creationTimestamp`,description="Age"
+// +kubebuilder:printcolumn:name="Programmed",type=string,JSONPath=`.status.conditions[?(@.type=="Programmed")].status`
 
 // KongConsumer is the Schema for the kongconsumers API.
 type KongConsumer struct {
@@ -45,6 +46,9 @@ type KongConsumer struct {
 	// Credentials are references to secrets containing a credential to be
 	// provisioned in Kong.
 	Credentials []string `json:"credentials,omitempty"`
+	// ConsumerGroups are references to consumer groups (that consumer wants to be part of)
+	// provisioned in Kong.
+	ConsumerGroups []string `json:"consumerGroups,omitempty"`
 
 	// Status represents the current status of the KongConsumer resource.
 	Status KongConsumerStatus `json:"status,omitempty"`