+++ CHANGELOG.md
@@ -76,7 +76,7 @@ Adding a new version? You'll need three changes:
 ### Added
 
 - **WIP** Introduce `KongConsumerGroup` CRD (supported by Kong Enterprise only)
-  [#4325](https://github.com/Kong/kubernetes-ingress-controller/pull/4325), [#4387](https://github.com/Kong/kubernetes-ingress-controller/pull/4387)
+  [#4325](https://github.com/Kong/kubernetes-ingress-controller/pull/4325), [#4387](https://github.com/Kong/kubernetes-ingress-controller/pull/4387), [#4419](https://github.com/Kong/kubernetes-ingress-controller/pull/4419)
 - The ResponseHeaderModifier Gateway API filter is now supported and translated
   to the proper set of Kong plugins.
   [#4350](https://github.com/Kong/kubernetes-ingress-controller/pull/4350)
@@ -128,11 +128,22 @@ Adding a new version? You'll need three changes:
   updates. The controller will now send configuration to such Gateways and will
   actively monitor their readiness for accepting configuration updates.
   [#4368](https://github.com/Kong/kubernetes-ingress-controller/pull/4368
-- `KongConsumer`, `KongPlugin`, and `KongClusterPlugin` CRDs were extended with
+- `KongConsumer`, `KongConsumerGroup` `KongPlugin`, and `KongClusterPlugin` CRDs were extended with
   `Status.Conditions` field. It will contain the `Programmed` condition describing
   whether an object was successfully translated into Kong entities and sent to Kong.
   [#4409](https://github.com/Kong/kubernetes-ingress-controller/pull/4409)
   [#4412](https://github.com/Kong/kubernetes-ingress-controller/pull/4412)
+  [#4423](https://github.com/Kong/kubernetes-ingress-controller/pull/4423)
+- `KongConsumer`, `KongConsumerGroup`, `KongPlugin`, and `KongClusterPlugin`'s `additionalPrinterColumns`
+  were extended with `Programmed` column. It will display the status of the
+  `Programmed` condition of an object when `kubectl get` is used.
+  [#4425](https://github.com/Kong/kubernetes-ingress-controller/pull/4425)
+  [#4423](https://github.com/Kong/kubernetes-ingress-controller/pull/4423)
+- Parser instead of logging errors for invalid `KongPlugin` or `KongClusterPlugin`
+  configuration, will now propagate a translation failure that will result
+  in the `Programmed` condition of the object being set to `False` and an
+  event being emitted.
+  [#4428](https://github.com/Kong/kubernetes-ingress-controller/pull/4428)
 
 ### Changed